requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
docker==3.6.0
docker-pycreds==0.4.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
multi-key-dict==2.0.3
pbr==5.1.1
pkg-resources==0.0.0
python-jenkins==1.4.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.54.0
Werkzeug==0.14.1
